package refactoring + script and style changes - User, Role classes moved to the domain.security package - browser history is now used in redirects
