+++ src/main/java/com/korotin/quizlet/repository/RoleRepository.java
@@ -1,6 +1,6 @@
 package com.korotin.quizlet.repository;
 
-import com.korotin.quizlet.domain.Role;
+import com.korotin.quizlet.domain.security.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
