+++ src/main/java/com/korotin/quizlet/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.korotin.quizlet.repository;
 
-import com.korotin.quizlet.domain.User;
+import com.korotin.quizlet.domain.security.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
