+++ src/main/resources/static/styles/login.css
@@ -68,6 +68,7 @@ body {
 }
 
 .submit-button:hover {
+    cursor: pointer;
     background-color: #89a5ff;
 }
